Move test data to a "testdata" package.
