+++ tests/routes/test_fitbit_routes.py
@@ -20,7 +20,7 @@ from tests.testsupport.factories.factories import (
     FitbitUserFactory,
     UserFactory,
 )
-from tests.testsupport.fixtures.fitbit_scenarios import (
+from tests.testsupport.testdata.fitbit_scenarios import (
     FitbitActivityScenario,
     FitbitSleepScenario,
     activity_scenarios,
