+++ tests/testsupport/testdata/fitbit_scenarios.py
