+++ tests/testsupport/testdata/fitbit_sleep_response_1_item.json
