+++ tests/testsupport/testdata/fitbit_sleep_response_classic.json
